refac: extract sort tasks, normalize functions and remove useLocalStorage
