--- frontend/src/hooks/useLocalStorage.js
@@ -1,23 +1,0 @@
-import { useState, useEffect } from "react";
-
-export const useLocalStorage = (key, initialValue) => {
-  const [storedValue, setStoredValue] = useState(() => {
-    try {
-      const item = localStorage.getItem(key);
-      return item ? JSON.parse(item) : initialValue;
-    } catch (error) {
-      console.error("Error reading from localStorage:", error);
-      return initialValue;
-    }
-  });
-
-  useEffect(() => {
-    try {
-      localStorage.setItem(key, JSON.stringify(storedValue));
-    } catch (error) {
-      console.error("Error writing to localStorage:", error);
-    }
-  }, [key, storedValue]);
-
-  return [storedValue, setStoredValue];
-};
